+++ frontend/src/api/admin/memberApi.ts
@@ -2,7 +2,7 @@ import api from "../axios";
 import { User } from "../../types/admin/user";
 
 // 사용자 관리 API
-export const memberApi = {
+const memberApi = {
   // 전체 사용자 목록 조회
   getMemberList: async (): Promise<User[]> => {
     const response = await api.get<User[]>("/admin/member");
@@ -18,3 +18,4 @@ export const memberApi = {
     return response.data;
   },
 };
+export default memberApi;
