+++ frontend/src/api/admin/productApi.ts
@@ -2,7 +2,7 @@ import api from "../axios";
 import { ProductListDto } from "../../types/admin/product";
 
 // 물품 관리 API
-export const productApi = {
+const productApi = {
   // 물품 목록 조회
   getProductList: () =>
     api.get<ProductListDto[]>("/admin/product/getProductList"),
@@ -44,3 +44,4 @@ export const productApi = {
       data: productId,
     }),
 };
+export default productApi;
