+++ frontend/src/api/index.ts
@@ -1,0 +1,5 @@
+export { default as dashboardApi } from "./admin/dashboardApi";
+export { default as productApi } from "./admin/productApi";
+export { default as travelApi } from "./admin/travelApi";
+export { default as memberApi } from "./admin/memberApi";
+export { default as orderApi } from "./admin/orderApi";
