+++ frontend/src/pages/admin/manageOrder/ManageOrder.tsx
@@ -3,7 +3,7 @@ import Header from "../../../components/common/Header_dash/Header";
 import Sidebar from "../../../components/common/Sidebar/Sidebar";
 import styles from "./ManageOrder.module.scss";
 import { FiSearch } from "react-icons/fi";
-import { orderApi } from "../../../api/admin/orderApi";
+import { orderApi } from "../../../api";
 import {
   getOrderStatusText,
   getOrderStatusClass,
